Fix Android build: stub mismatched P2P/E2EE/WebRTC libs, fix Hilt metadata, clean root Gradle, remove committed build artifacts
Changes:

- Root build.gradle.kts reduced to no-op; removed :ios placeholder from settings.gradle.kts

- Hilt 2.52 -> 2.56.1 for Kotlin 2.1 metadata compatibility

- LibP2PManager, SignalProtocol, WebRTCManager stubbed due to dependency/API mismatch

- ReputationSystem Wilson score fixed for Int overflow (Long total)

- AppDatabase/DI/DAOs/entities aligned; removed unused Signal stores

- NostrClientTest and ReputationSystemTest fixed and passing

- Removed committed android/.gradle and android/app/build artifacts; updated .gitignore

+++ android/app/src/test/java/com/neop2p/data/reputation/ReputationSystemTest.kt
@@ -16,8 +16,8 @@ class ReputationSystemTest {
      * Replicate the Wilson score calculation for testability.
      */
     private fun calculateWilsonScore(positive: Int, negative: Int): Float {
-        val total = positive + negative
-        if (total == 0) return 0f
+        val total = positive.toLong() + negative.toLong()
+        if (total == 0L) return 0f
         val z = 1.96
         val p = positive.toDouble() / total
         val left = p + (z * z) / (2 * total)
